Move SNMP show commands to separated file
